added exportable textbox story props
