+++ src/__experimental__/components/textbox/textbox.stories.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { boolean, text, select } from '@storybook/addon-knobs';
+import {
+  boolean,
+  text,
+  select,
+  number
+} from '@storybook/addon-knobs';
 import Textbox from '.';
 import OptionsHelper from '../../../utils/helpers/options-helper';
 
@@ -8,52 +13,70 @@ storiesOf('Experimental/Textbox', module)
   .add('Basic', () => {
     return (
       <Textbox
-        disabled={ boolean('disabled') }
         error={ text('error') }
-        fieldHelp={ text('fieldHelp') }
         info={ text('info') }
-        label={ text('label') }
-        labelAlign={ select('labelAlign', OptionsHelper.alignBinary) }
-        labelHelp={ text('labelHelp') }
-        labelInline={ boolean('labelInline') }
-        placeholder={ text('placeholder') }
-        readOnly={ boolean('readOnly') }
         size={ select('size', OptionsHelper.sizesRestricted) }
         warning={ text('warning') }
+        { ...getTextboxStoryProps() }
       />
     );
   })
   .add('Multiple', () => {
     return ([
       <Textbox
-        disabled={ boolean('disabled') }
         error={ text('error') }
-        fieldHelp={ text('fieldHelp') }
         info={ text('info') }
-        label={ text('label') }
-        labelAlign={ select('labelAlign', OptionsHelper.alignBinary) }
-        labelHelp={ text('labelHelp') }
-        labelInline={ boolean('labelInline') }
-        placeholder={ text('placeholder') }
-        readOnly={ boolean('readOnly') }
         size={ select('size', OptionsHelper.sizesRestricted) }
         warning={ text('warning') }
         key='0'
+        { ...getTextboxStoryProps() }
       />,
       <Textbox
-        disabled={ boolean('disabled') }
         error={ text('error') }
-        fieldHelp={ text('fieldHelp') }
         info={ text('info') }
-        label={ text('label') }
-        labelAlign={ select('labelAlign', OptionsHelper.alignBinary) }
-        labelHelp={ text('labelHelp') }
-        labelInline={ boolean('labelInline') }
-        placeholder={ text('placeholder') }
-        readOnly={ boolean('readOnly') }
         size={ select('size', OptionsHelper.sizesRestricted) }
         warning={ text('warning') }
         key='1'
+        { ...getTextboxStoryProps() }
       />
     ]);
   });
+
+
+function getTextboxStoryProps() {
+  const percentageRange = {
+    range: true,
+    min: 0,
+    max: 100,
+    step: 1
+  };
+  const disabled = boolean('disabled', false);
+  const readOnly = boolean('readOnly', false);
+  const placeholder = text('placeholder');
+  const fieldHelp = text('fieldHelp', '');
+  const label = text('label', '');
+  const labelHelp = label ? text('labelHelp') : undefined;
+  const labelInline = label ? select('labelInline', OptionsHelper.alignBinary) : undefined;
+  const labelWidth = labelInline ? number('labelWidth', 30, percentageRange) : undefined;
+  const inputWidth = labelInline ? number('inputWidth', 70, percentageRange) : undefined;
+  const labelAlign = labelInline ? select(
+    'labelAlign',
+    OptionsHelper.alignBinary,
+    OptionsHelper.alignBinary[0]
+  ) : undefined;
+
+  return {
+    disabled,
+    readOnly,
+    placeholder,
+    inputWidth,
+    fieldHelp,
+    label,
+    labelHelp,
+    labelInline,
+    labelWidth,
+    labelAlign
+  };
+}
+
+export default getTextboxStoryProps;
